Allow build of tbot without cgo
