+++ lib/auth/keystore/pkcs11_no_cgo.go
@@ -1,0 +1,28 @@
+//go:build !cgo
+
+/*
+Copyright 2023 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package keystore
+
+import (
+	"github.com/gravitational/trace"
+	"github.com/sirupsen/logrus"
+)
+
+func newPKCS11KeyStore(config any, logger logrus.FieldLogger) (backend, error) {
+	return nil, trace.NotImplemented("pkcs11 requires cgo")
+}
