+++ lib/inventory/metadata/metadata_linux_cgo_test.go
